deployment_manager/requirements.txt
azure_storage_blob
docker==6.0.1
filelock==3.8.0
Flask
kafka-python==2.0.2
packaging==21.3
paramiko==3.1.0
protobuf==4.22.1
psutil==5.9.4
pyftpdlib==1.5.7
PyJWT
pymongo==4.3.3
python-ftp-server==1.3.17
requests
service-identity==21.1.0
shortuuid==1.0.11
threadpoolctl==3.1.0
tqdm==4.65.0
virtualenv==20.16.6